+++ src/main/java/com/fastcampaus/springpractice/logic/BubbleSort.java
@@ -1,5 +1,7 @@
-package com.fastcampaus.javaoop.logic;
+package com.fastcampaus.springpractice.logic;
 
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -7,6 +9,7 @@ import java.util.List;
 // 1.여러가지 자료형을 정렬할수 있는 서비스
 // 2. 정렬할수 있는다는거 비교가 가능해야 한다는거 (제네릭을 사용)
 //  <T extends Comparable<T>> --> T(제네릭, 즉 비교대상) 은 항상 Comparable 을 상속하는 자료형만 받을수있다.
+@Component // .해당 어노테이션으로 BubbleSort가 빈으로 인식됨
 public class BubbleSort <T extends Comparable<T>> implements Sort<T>{
     @Override
     public List<T> sort(List<T> list) {
