+++ src/main/java/com/fastcampaus/springpractice/logic/JavaSort.java
@@ -1,9 +1,12 @@
-package com.fastcampaus.javaoop.logic;
+package com.fastcampaus.springpractice.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
+@Component
 public class JavaSort <T extends Comparable<T>> implements Sort<T>{
     @Override
     public List<T> sort(List<T> list) {
