+++ src/test/java/com/fastcampaus/springpractice/MainTest.java
@@ -1,8 +1,6 @@
-package com.fastcampaus.javaoop;
+package com.fastcampaus.springpractice;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class MainTest {
